Revert "[1주차] 객체지향 코드 연습(Hoya324) (#1)"
This reverts commit 53db6a6d6fb03876a1a670582382acd2d5a76129.
