--- docs/README.md
@@ -1,58 +1,0 @@
-# 구현해야할 목록
-
-- 입출력
-- 입출력과 데이터 저장, 로직구현을 연결짓기
-- 로직구현
-- 데이터 저장
-- 에러처리
-
-## 입출력(View)
-
-### Lotto 메세지 입력 
-
-- 구입금액을 입력 받아서 로또 구매량을 반환 메서드
-  - 정수값인지 확인
-- 당첨번호 입력 받아서 List 형태로 반환하는 메서드
-- 보너스 번호를 입력 받아서 당첨번호 목록에 저장한 List를 반환하는 메서드
-
-### Lotto 메시지 출력
-
-- 구매한 로또의 수만큼 로또의 번호를 저장하여 다시 출력하는 메서드
-- 구매한 로또 번호와 당첨 로또 번호를 비교하여 그 결과(당첨통계)를 출력하는 메서드
-  - 폼에 맞게 문자열을 변환하는 메서드
-  - 당첨금을 계산하는 메서드
-- 수익률을 계산한는 메서드
-
-## 입출력과 데이터 저장, 로직구현을 연결짓기(Controller)
-
-## 로직구현(Service)
-- 로또 구매 수를 계산해주는 메서드
-- 로또를 랜덤의 수로 생성하는 메서드
-- List 형태로 정제되지 않은 당첨번호를 List로 처리하는 메서드
-  - 해당 수가 1~45 사이의 수인지 판단
-- 보너스 번호를 저장하는 메서드
-  - 해당 수가 1~45 사이의 수인지 판단
-- 당첨금액을 계산하는 메서드
-- 당첨 결과를 등수에 맞게 저장하는 메서드
-
-## 데이터 저장(Model)
-
-### 사용자가 구매한 로또 저장
-- 구입한 Lotto의 수만큼 저장하는 메서드
-- 저장된 값을 출력하는 메서드
-
-### 당첨 번호 저장
-- 당첨 번호를 저장하는 메서드
-- 보너스를 저장하는 메서드
-- 저장된 값을 출력하는 메서드
-
-### 당첨 번호와 구매한 로또를 비교하고 그 결과를 저장
-- 최소 3개 이상 맞춘 경우 해당하는 당첨 결과 배열의 인덱스에 기록하는 메서드
-- 저장된 값을 출력하는 메서드
-
-## 에러처리
-- 로또번호는 1~45 사이에 속하는 값이다.
-- 구매금액은 1000의 자리 수이다.
-- 구매금액은 1000 이상의 수이다.
-- 로또 번호는 6개이다.
-- 각각의 로또 번호는 중복되지 않는다.
