--- src/main/java/lotto/model/LottoNumberRepository.java
@@ -1,12 +1,0 @@
-package lotto.model;
-
-import lotto.domain.Lotto;
-
-import java.util.List;
-
-public interface LottoNumberRepository {
-
-  void saveLotto(Lotto lotto);
-
-  List<Lotto> findLottos();
-}
