--- src/main/java/lotto/model/LottoResultRepository.java
@@ -1,10 +1,0 @@
-package lotto.model;
-
-import java.util.List;
-
-public interface LottoResultRepository {
-
-  void saveLottoResult(int winningCount, boolean isBonus);
-
-  List<Integer> findLottoResults();
-}
